Move `XContent` -> `SnapshotInfo` parsing out of prod
The code to parse a `SnapshotInfo` object out of an `XContent` response
body is only used in tests, so this commit moves it out of the
production codebase and into the test framework.
