+++ server/src/main/java/org/elasticsearch/action/admin/cluster/snapshots/create/CreateSnapshotResponse.java
@@ -14,12 +14,8 @@ import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.core.Nullable;
 import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.snapshots.SnapshotInfo;
-import org.elasticsearch.snapshots.SnapshotInfo.SnapshotInfoBuilder;
-import org.elasticsearch.xcontent.ObjectParser;
-import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContentObject;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -29,24 +25,8 @@ import java.util.Objects;
  */
 public class CreateSnapshotResponse extends ActionResponse implements ToXContentObject {
 
-    private static final ObjectParser<CreateSnapshotResponse, Void> PARSER = new ObjectParser<>(
-        CreateSnapshotResponse.class.getName(),
-        true,
-        CreateSnapshotResponse::new
-    );
-
-    static {
-        PARSER.declareObject(
-            CreateSnapshotResponse::setSnapshotInfoFromBuilder,
-            SnapshotInfo.SNAPSHOT_INFO_PARSER,
-            new ParseField("snapshot")
-        );
-    }
-
     @Nullable
-    private SnapshotInfo snapshotInfo;
-
-    CreateSnapshotResponse() {}
+    private final SnapshotInfo snapshotInfo;
 
     public CreateSnapshotResponse(@Nullable SnapshotInfo snapshotInfo) {
         this.snapshotInfo = snapshotInfo;
@@ -55,10 +35,6 @@ public class CreateSnapshotResponse extends ActionResponse implements ToXContent
     public CreateSnapshotResponse(StreamInput in) throws IOException {
         super(in);
         snapshotInfo = in.readOptionalWriteable(SnapshotInfo::readFrom);
-    }
-
-    private void setSnapshotInfoFromBuilder(SnapshotInfoBuilder snapshotInfoBuilder) {
-        this.snapshotInfo = snapshotInfoBuilder.build();
     }
 
     /**
@@ -101,10 +77,6 @@ public class CreateSnapshotResponse extends ActionResponse implements ToXContent
         }
         builder.endObject();
         return builder;
-    }
-
-    public static CreateSnapshotResponse fromXContent(XContentParser parser) {
-        return PARSER.apply(parser, null);
     }
 
     @Override
